# faster move gen by replacing hash map with [[[bool; 4]; 25]; 10]