@MinusKelvin 2022-04-12
There are a number of issues with this PR:
1. You changed a *lot* more than just replacing the hash map with a `[[[bool; 4]; 25]; 10]`.
2. The code does not compile (assigns twice to immutable variable, line 202).
3. I don't have to look very hard at your rewrite to know that it does not count the minimum number of softdrops to place in a location correctly.
4. I tested this with `battletris` using game config `ppt@2` as an elo gainer: `1352 - 1536   LLR: -2.96 (-2.94, 2.94)  Elo: -22.17 ± 12.69`
5. A `[[[bool; 4]; 25]; 10]` is not large enough in general (tetris boards are 40 rows tall), and one game in the above match was forfeit due to a crash as a result of this.

That being said, there is likely some performance to be gained here by abandoning precise softdrop counting in favor of an approximation, and then replacing the hashmaps and priority queue with alternative datastructures.